+++ ios/Classes/core/PMAssetPathEntity.h
@@ -12,7 +12,7 @@
 @property(nonatomic, assign) BOOL isAll;
 @property(nonatomic, assign) int type;
 @property(nonatomic, assign) NSUInteger assetCount;
-@property(nonatomic, assign) long modifiedDate;
+@property(nonatomic, assign) double modifiedDate;
 @property(nonatomic, strong) PHAssetCollection *collection;
 
 + (instancetype)entityWithId:(NSString *)id name:(NSString *)name assetCollection:(PHAssetCollection*)collection;
@@ -22,13 +22,13 @@
 @interface PMAssetEntity : NSObject
 
 @property(nonatomic, copy) NSString *id;
-@property(nonatomic, assign) long createDt;
+@property(nonatomic, assign) double createDt;
 @property(nonatomic, assign) NSUInteger width;
 @property(nonatomic, assign) NSUInteger height;
-@property(nonatomic, assign) long duration;
+@property(nonatomic, assign) double duration;
 @property(nonatomic, assign) int type;
 @property(nonatomic, strong) PHAsset *phAsset;
-@property(nonatomic, assign) long modifiedDt;
+@property(nonatomic, assign) double modifiedDt;
 @property(nonatomic, assign) double lat;
 @property(nonatomic, assign) double lng;
 @property(nonatomic, copy) NSString *title;
@@ -37,17 +37,17 @@
 @property(nonatomic, assign) BOOL isLocallyAvailable;
 
 - (instancetype)initWithId:(NSString *)id
-                  createDt:(long)createDt
+                  createDt:(double)createDt
                      width:(NSUInteger)width
                     height:(NSUInteger)height
-                  duration:(long)duration
+                  duration:(double)duration
                       type:(int)type;
 
 + (instancetype)entityWithId:(NSString *)id
-                    createDt:(long)createDt
+                    createDt:(double)createDt
                        width:(NSUInteger)width
                       height:(NSUInteger)height
-                    duration:(long)duration
+                    duration:(double)duration
                         type:(int)type;
 
 @end
